Split mockserver.py.
