+++ tests/test_closespider.py
@@ -1,7 +1,7 @@
 from twisted.internet.defer import inlineCallbacks
 
 from scrapy.utils.test import get_crawler
-from tests.mockserver import MockServer
+from tests.mockserver.http import MockServer
 from tests.spiders import (
     ErrorSpider,
     FollowAllSpider,
