+++ tests/test_command_fetch.py
@@ -1,4 +1,4 @@
-from tests.mockserver import MockServer
+from tests.mockserver.http import MockServer
 from tests.test_commands import TestProjectBase
