+++ tests/test_pipelines.py
@@ -6,7 +6,7 @@ from twisted.internet.defer import Deferred, inlineCallbacks
 from scrapy import Request, Spider, signals
 from scrapy.utils.defer import deferred_to_future, maybe_deferred_to_future
 from scrapy.utils.test import get_crawler, get_from_asyncio_queue
-from tests.mockserver import MockServer
+from tests.mockserver.http import MockServer
 
 
 class SimplePipeline:
