+++ tests/test_proxy_connect.py
@@ -12,7 +12,7 @@ from twisted.internet.defer import inlineCallbacks
 
 from scrapy.http import Request
 from scrapy.utils.test import get_crawler
-from tests.mockserver import MockServer
+from tests.mockserver.http import MockServer
 from tests.spiders import SimpleSpider, SingleRequestSpider
