+++ tests/test_spidermiddleware_output_chain.py
@@ -3,7 +3,7 @@ from testfixtures import LogCapture
 from scrapy import Request, Spider
 from scrapy.utils.defer import deferred_f_from_coro_f, maybe_deferred_to_future
 from scrapy.utils.test import get_crawler
-from tests.mockserver import MockServer
+from tests.mockserver.http import MockServer
 
 
 class LogExceptionMiddleware:
