+++ tests/utils/__init__.py
@@ -1,3 +1,6 @@
+import os
+from pathlib import Path
+
 from twisted.internet.defer import Deferred
 
 
@@ -7,3 +10,13 @@ def twisted_sleep(seconds):
     d = Deferred()
     reactor.callLater(seconds, d.callback, None)
     return d
+
+
+def get_script_run_env() -> dict[str, str]:
+    """Return a OS environment dict suitable to run scripts shipped with tests."""
+
+    tests_path = Path(__file__).parent.parent
+    pythonpath = str(tests_path) + os.pathsep + os.environ.get("PYTHONPATH", "")
+    env = os.environ.copy()
+    env["PYTHONPATH"] = pythonpath
+    return env
